requirements.txt
pystray
pillow
flask 
